# packages/types/openapi.yaml
openapi: 3.1.0
info:
  title: Mirage API
  version: 0.0.0
  description: |
    The single OpenAPI spec for Mirage. Lives in `@mirage/types` so it can
    be consumed both by the BFF (for route shape) and the React SPA (for the
    typed client). Generated TypeScript types land in `openapi.generated.ts`
    via `pnpm gen:openapi`, which is wired to `postinstall`.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
paths:
...
  /workspaces:
    get:
      summary: List the current org's Workspaces
      operationId: listWorkspaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a Workspace
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workspaces/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Get a Workspace by id
      operationId: getWorkspace
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a Workspace (cascade)
      description: |
        Soft-deletes the workspace immediately (sets `deletedAt`) and starts
        an async cascade that cancels in-flight runs, purges every S3 object
        under `org/{orgId}/workspace/{wsId}/`, and deletes every Mongo doc
        scoped to this workspace. Caller must be an org owner.
      operationId: deleteWorkspace
      responses:
        '202':
          description: Accepted — soft delete written; cascade in progress
          content:
            application/json:
              schema:
                type: object
                required: [id, deletedAt]
                additionalProperties: false
                properties:
                  id: { type: string }
                  deletedAt: { type: string, format: date-time }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Workspace is already being deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /workspaces/{wsId}/sets:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
    get:
      summary: List Sets in a Workspace
      operationId: listSets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Set' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Create a Set in a Workspace
      operationId: createSet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSetBody' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Set' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{wsId}/sets/{id}:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Get a Set by id
      operationId: getSet
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Set' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update a Set
      operationId: updateSet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSetBody' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Set' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: Delete a Set
      operationId: deleteSet
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{wsId}/sets/{id}/edges:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Computed cross-schema reference edges for this Set
      operationId: getSetEdges
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SetEdge' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workspace:
      type: object
      required: [id, orgId, name, createdBy, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string }
        orgId: { type: string }
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt:
          type: string
          format: date-time
          description: |
            Set when the workspace is soft-deleted. While present, the row is
            still returned by list/get but every sub-resource endpoint
            responds 404. The row is hard-deleted (removed entirely) once the
            background cascade finishes.
    CreateWorkspaceBody:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
...
    Strategy:
      oneOf:
        - type: object
          required: [type]
          additionalProperties: false
          properties:
            type: { type: string, enum: ['1:1'] }
        - type: object
          required: [type]
          additionalProperties: false
          properties:
            type: { type: string, enum: ['random'] }
            allowDuplicates: { type: boolean }
        - type: object
          required: [type]
          additionalProperties: false
          properties:
            type: { type: string, enum: ['evenSplit'] }
        - type: object
          required: [type, functionId]
          additionalProperties: false
          properties:
            type: { type: string, enum: ['custom'] }
            functionId: { type: string }
    StrategyOverride:
      type: object
      required: [schemaKey, fieldPath, strategy]
      additionalProperties: false
      properties:
        schemaKey: { type: string }
        fieldPath: { type: string }
        strategy: { $ref: '#/components/schemas/Strategy' }
    SetSchemaInclusion:
      type: object
      required: [schemaKey, count]
      additionalProperties: false
      properties:
        schemaKey: { type: string }
        count: { type: integer, minimum: 0, maximum: 10000000 }
    SetOutputConfig:
      type: object
      required: [format, locale, workerPool]
      additionalProperties: false
      properties:
        format: { type: string, enum: [json, ndjson, csv, sql, parquet] }
        locale: { type: string, minLength: 2, maxLength: 16 }
        workerPool: { type: integer, minimum: 1, maximum: 64 }
    Set:
      type: object
      required:
        - id
        - workspaceId
        - orgId
        - key
        - name
        - description
        - color
        - icon
        - tags
        - salt
        - schemas
        - strategies
        - output
        - createdBy
        - createdAt
        - updatedAt
      additionalProperties: false
      properties:
        id: { type: string }
        workspaceId: { type: string }
        orgId: { type: string }
        key:
          type: string
          pattern: '^[a-z][a-z0-9-]{0,39}$'
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        color:
          type: string
          enum: [violet, cyan, emerald, amber, rose, slate]
        icon: { type: string }
        tags:
          type: array
          items: { type: string }
        salt: { type: string, minLength: 1, maxLength: 64 }
        schemas:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SetSchemaInclusion' }
        strategies:
          type: array
          items: { $ref: '#/components/schemas/StrategyOverride' }
        output: { $ref: '#/components/schemas/SetOutputConfig' }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateSetBody:
      type: object
      required: [key, name, color, icon, tags, salt, schemas, strategies, output]
      additionalProperties: false
      properties:
        key:
          type: string
          pattern: '^[a-z][a-z0-9-]{0,39}$'
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        color:
          type: string
          enum: [violet, cyan, emerald, amber, rose, slate]
        icon: { type: string }
        tags:
          type: array
          items: { type: string }
        salt: { type: string, minLength: 1, maxLength: 64 }
        schemas:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SetSchemaInclusion' }
        strategies:
          type: array
          items: { $ref: '#/components/schemas/StrategyOverride' }
        output: { $ref: '#/components/schemas/SetOutputConfig' }
    UpdateSetBody:
      type: object
      required:
        - key
        - name
        - color
        - icon
        - tags
        - salt
        - schemas
        - strategies
        - output
        - expectedUpdatedAt
      additionalProperties: false
      properties:
        key:
          type: string
          pattern: '^[a-z][a-z0-9-]{0,39}$'
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        color:
          type: string
          enum: [violet, cyan, emerald, amber, rose, slate]
        icon: { type: string }
        tags:
          type: array
          items: { type: string }
        salt: { type: string, minLength: 1, maxLength: 64 }
        schemas:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SetSchemaInclusion' }
        strategies:
          type: array
          items: { $ref: '#/components/schemas/StrategyOverride' }
        output: { $ref: '#/components/schemas/SetOutputConfig' }
        expectedUpdatedAt: { type: string, format: date-time }
    SetEdge:
      type: object
      required: [fromSchemaKey, fromFieldPath, toSchemaKey, cardinality]
      additionalProperties: false
      properties:
        fromSchemaKey: { type: string }
        fromFieldPath: { type: string }
        toSchemaKey: { type: string }
        cardinality:
          type: string
          enum: [one, many]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        detail: {}
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }